requirements.txt
urllib3==1.26.14
kafka-python==2.0.2
redis==4.5.4